Removed AccountValidationButtons.js because phone button is unneeded. Verify Email button moved to UserPage.js
